Creation of auth service ! + Added new proto file for auth + Did the necessary stuff on client-side + Moved the server test functions + Moved keycloak functions to common + Regenerated proto for client + Modified envoy config to add a service + Modified compose files + Auth gRPC runs on port 50054
