+++ docker-compose.dev.yml
@@ -18,4 +18,5 @@ services:
       - 50051:50051
       - 50052:50052
       - 50053:50053
+      - 50054:50054
