+++ generateProtos.sh
@@ -4,3 +4,4 @@
 protoc protofiles/user.proto --go_out=plugins=grpc:./server/
 protoc protofiles/chatroom.proto --go_out=plugins=grpc:./server/
 protoc protofiles/message.proto --go_out=plugins=grpc:./server/
+protoc protofiles/auth.proto --go_out=plugins=grpc:./server/
