+++ owly-client/src/_services/auth.service.ts
@@ -2,8 +2,9 @@ import { Injectable } from '@angular/core';
 import { JwtHelperService } from '@auth0/angular-jwt';
 import { CookieService } from 'ngx-cookie';
 import { BehaviorSubject } from 'rxjs';
+import { AuthServiceClient } from 'src/proto/auth.pbsc';
 import { Chatroom } from 'src/proto/chatroom.pb';
-import { LoginUserRequest } from 'src/proto/user.pb';
+import { CreateNewUserRequest, LoginUserRequest } from 'src/proto/auth.pb';
 import { UserServiceClient } from 'src/proto/user.pbsc';
 import { LoggedUser } from 'src/_models/loggedUser';
 import { ChatroomService } from './chatroom.service';
@@ -17,7 +18,7 @@ export class AuthService {
     constructor(
         public jwtHelper: JwtHelperService,
         private cookieService: CookieService, 
-        
+        private authClient: AuthServiceClient,
         private userClient: UserServiceClient,
         private storeService: StoreService) { 
         const user:LoggedUser = <LoggedUser>this.cookieService.getObject("owly_user_cookies");
@@ -41,6 +42,23 @@ export class AuthService {
         return !this.jwtHelper.isTokenExpired(token);
     }
 
+    createUser(data) {
+        const request = this.createUserRequest(data);
+        this.authClient.createNewUser(request).subscribe(
+            (res) => console.log(res),
+            (err) => console.log(err)
+        );
+    }
+
+    private createUserRequest(data) {
+        return new CreateNewUserRequest({
+            firstName: data.firstName,
+            lastName: data.lastName,
+            username: data.username,
+            email: data.email,
+            password: data.password
+        })
+    }
 
     createLoggedUserFromJWT(jwt: string): LoggedUser {
         var infos = this.jwtHelper.decodeToken(jwt);
@@ -51,7 +69,7 @@ export class AuthService {
 
 
     async login(req: LoginUserRequest): Promise<Boolean> {
-        const res = await this.userClient.loginUser(req).toPromise();
+        const res = await this.authClient.loginUser(req).toPromise();
         if (res.result.accessToken == "") {
             return false;
         }
