+++ owly-client/src/_services/user.service.ts
@@ -3,51 +3,36 @@ import { GRPC_CLIENT_FACTORY } from '@ngx-grpc/core';
 import { GrpcWebClientBase } from 'grpc-web';
 import { GrpcClientSettings, GrpcStatusEvent } from '@ngx-grpc/common';
 import { GRPC_USER_SERVICE_CLIENT_SETTINGS } from 'src/proto/user.pbconf';
-import { CreateNewUserRequest, CreateNewUserResponse, JWT, LoginUserRequest, SearchUserByUsernameRequest, User } from 'src/proto/user.pb';
+
 import { UserServiceClient } from 'src/proto/user.pbsc';
 import { LoggedUser } from 'src/_models/loggedUser';
 import { BehaviorSubject } from 'rxjs';
 import { CookieService } from 'ngx-cookie';
 import { AuthService } from './auth.service';
+import { SearchUserByUsernameRequest, User } from 'src/proto/user.pb';
+import { AuthServiceClient } from 'src/proto/auth.pbsc';
 
-interface CreateNewUserEvent {
-    request?: CreateNewUserRequest,
-    response?: CreateNewUserResponse,
-    status?: GrpcStatusEvent,
-    error?: GrpcStatusEvent
-}
 
 
 @Injectable({ providedIn: 'root' })
 export class UserService {
 
+    currentUser: LoggedUser;
+
     constructor(
-        private userClient: UserServiceClient
-    ) { 
-        
-    }
+        private userClient: UserServiceClient,
+        private authService: AuthService
+    ) {
+        this.authService.currentUser.subscribe((user) => {
+            this.currentUser = user;
+          });
+     }
 
-
-    createUser(data) {
-        const request = this.createUserRequest(data);
-        this.userClient.createNewUser(request).subscribe(
-            (res) => console.log(res),
-            (err) => console.log(err)
-        );
-    }
-
-    private createUserRequest(data) {
-        return new CreateNewUserRequest({
-            firstName: data.firstName,
-            lastName: data.lastName,
-            username: data.username,
-            email: data.email,
-            password: data.password
-        })
-    }
 
     async searchUserByUsername(req: SearchUserByUsernameRequest): Promise<User[]> {
-        const res = await this.userClient.searchUserByUsername(req).toPromise();
+        const token = this.currentUser.accessToken;
+
+        const res = await this.userClient.searchUserByUsername(req, {"authorization": token}).toPromise();
         return res.results;
     }
