+++ owly-client/src/app/app.module.ts
@@ -24,6 +24,7 @@ import { CookieModule } from 'ngx-cookie';
 import { TextareaAutosizeModule } from 'ngx-textarea-autosize';
 import { GRPC_MESSAGE_SERVICE_CLIENT_SETTINGS } from 'src/proto/message.pbconf';
 import { ChatroomListComponent } from './navigation/chatroom-list/chatroom-list.component';
+import { GRPC_AUTH_SERVICE_CLIENT_SETTINGS } from 'src/proto/auth.pbconf';
 
 @NgModule({
   declarations: [
@@ -52,7 +53,8 @@ import { ChatroomListComponent } from './navigation/chatroom-list/chatroom-list.
     { provide: GRPC_CLIENT_FACTORY, useClass: GrpcWebClientFactory },
     { provide: GRPC_USER_SERVICE_CLIENT_SETTINGS, useValue: { host: 'http://localhost:8085' } },
     {provide: GRPC_CHATROOM_SERVICE_CLIENT_SETTINGS, useValue: {host: 'http://localhost:8085'}},
-    { provide: GRPC_MESSAGE_SERVICE_CLIENT_SETTINGS, useValue: {host: 'http://localhost:8085'}}
+    { provide: GRPC_MESSAGE_SERVICE_CLIENT_SETTINGS, useValue: {host: 'http://localhost:8085'}},
+    { provide: GRPC_AUTH_SERVICE_CLIENT_SETTINGS, useValue: {host: 'http://localhost:8085'}},
   ],
   bootstrap: [AppComponent]
 })
