+++ owly-client/src/proto/auth.pbconf.ts
@@ -1,0 +1,15 @@
+/* tslint:disable */
+/* eslint-disable */
+//
+// THIS IS A GENERATED FILE
+// DO NOT MODIFY IT! YOUR CHANGES WILL BE LOST
+import { InjectionToken } from '@angular/core';
+import { GrpcClientSettings } from '@ngx-grpc/common';
+
+/**
+ * Specific GrpcClientSettings for AuthService.
+ * Use it only if your default settings are not set or the service requires other settings.
+ */
+export const GRPC_AUTH_SERVICE_CLIENT_SETTINGS = new InjectionToken<
+  GrpcClientSettings
+>('GRPC_AUTH_SERVICE_CLIENT_SETTINGS');
