+++ protofiles/user.proto
@@ -13,41 +13,10 @@ message User {
     string email = 3;
 }
 
-message JWT {
-    string AccessToken = 1;
-    string IDToken = 2;
-    int64 ExpiresIn = 3;
-    int64 RefreshExpiresIn = 4;
-    string RefreshToken = 5;
-    string TokenType = 6;
-    int64 NotBeforePolicy = 7;
-    string SessionState = 8;
-    string Scope = 9;
-}
 
-message CreateNewUserRequest {
-    string first_name = 1;
-    string last_name = 2;
-    string username = 3;
-    string email = 4;
-    string password = 5;
-}
-
-message CreateNewUserResponse {
-    bool success = 1;
-}
 
 message SearchUserByUsernameRequest {
     string username = 1;
-}
-
-message LoginUserRequest {
-    string username = 1;
-    string password = 2;
-}
-
-message LoginUserResponse {
-    JWT result = 1;
 }
 
 message SearchUserByUsernameResponse {
@@ -58,9 +27,6 @@ message SearchUserByUsernameResponse {
 
 
 service UserService {
-
-    rpc CreateNewUser(CreateNewUserRequest) returns(CreateNewUserResponse) {}
-    rpc LoginUser(LoginUserRequest) returns(LoginUserResponse) {}
 
     rpc SearchUserByUsername(SearchUserByUsernameRequest) returns(SearchUserByUsernameResponse) {}
 }
