+++ server/common/keycloak/admin.go
@@ -1,4 +1,4 @@
-package keycloak
+package common_keycloak
 
 import (
 	"context"
