+++ server/main.go
@@ -2,6 +2,7 @@ package main
 
 import (
 	"log"
+	authserver "primitivofr/owly/server/auth/auth_server"
 	"primitivofr/owly/server/chatroom/chatroom_server"
 	"primitivofr/owly/server/message/message_server"
 	"primitivofr/owly/server/user/user_server"
@@ -16,5 +17,6 @@ func main() {
 
 	go chatroom_server.StartServer()
 	go message_server.StartServer()
+	go authserver.StartServer()
 	user_server.StartServer()
 }
