+++ server/runTests.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+
+cd ./auth/auth_server && go test
